Update run.sh to allow stdin/stdout pipelining
The README.md assumes that you call run.sh with input on stdin, but the bash script assumed the filename to be given as argument. The new version makes a temp file for the input so it can be piped to both the feature extraction and the naf generation. 

I've also added a $DIR/ prefix for all calls so you can call it from another location if desired

(Note: this code is mostly written by @rubenIzquierdo)